dnfdaemon: move run_in_thread() to ThreadsManager
