+++ dnfdaemon-server/services/goal/goal.cpp
@@ -39,11 +39,11 @@ void Goal::dbus_register() {
     auto dbus_object = session.get_dbus_object();
     dbus_object->registerMethod(
         dnfdaemon::INTERFACE_GOAL, "resolve", "a{sv}", "a(ua{sv})", [this](sdbus::MethodCall call) -> void {
-            session.run_in_thread(*this, &Goal::resolve, std::move(call));
+            session.get_threads_manager().run_in_thread(*this, &Goal::resolve, std::move(call));
         });
     dbus_object->registerMethod(
         dnfdaemon::INTERFACE_GOAL, "do_transaction", "a{sv}", "", [this](sdbus::MethodCall call) -> void {
-            session.run_in_thread(*this, &Goal::do_transaction, std::move(call));
+            session.get_threads_manager().run_in_thread(*this, &Goal::do_transaction, std::move(call));
         });
 }
