+++ dnfdaemon-server/services/repoconf/repo_conf.cpp
@@ -31,19 +31,19 @@ void RepoConf::dbus_register() {
     auto dbus_object = session.get_dbus_object();
     dbus_object->registerMethod(
         dnfdaemon::INTERFACE_REPOCONF, "list", "a{sv}", "aa{sv}", [this](sdbus::MethodCall call) -> void {
-            session.run_in_thread(*this, &RepoConf::list, std::move(call));
+            session.get_threads_manager().run_in_thread(*this, &RepoConf::list, std::move(call));
         });
     dbus_object->registerMethod(
         dnfdaemon::INTERFACE_REPOCONF, "get", "s", "a{sv}", [this](sdbus::MethodCall call) -> void {
-            session.run_in_thread(*this, &RepoConf::get, std::move(call));
+            session.get_threads_manager().run_in_thread(*this, &RepoConf::get, std::move(call));
         });
     dbus_object->registerMethod(
         dnfdaemon::INTERFACE_REPOCONF, "enable", "as", "as", [this](sdbus::MethodCall call) -> void {
-            session.run_in_thread(*this, &RepoConf::enable, std::move(call));
+            session.get_threads_manager().run_in_thread(*this, &RepoConf::enable, std::move(call));
         });
     dbus_object->registerMethod(
         dnfdaemon::INTERFACE_REPOCONF, "disable", "as", "as", [this](sdbus::MethodCall call) -> void {
-            session.run_in_thread(*this, &RepoConf::disable, std::move(call));
+            session.get_threads_manager().run_in_thread(*this, &RepoConf::disable, std::move(call));
         });
 }
