+++ dnfdaemon-server/session.hpp
@@ -82,28 +82,6 @@ public:
     void fill_sack();
     bool read_all_repos();
 
-    template <class S>
-    void run_in_thread(S & service, sdbus::MethodReply (S::*method)(sdbus::MethodCall &&), sdbus::MethodCall && call) {
-        auto worker = std::thread(
-            [&method, &service, this](sdbus::MethodCall call) {
-                try {
-                    auto reply = (service.*method)(std::move(call));
-                    reply.send();
-                } catch (std::exception & ex) {
-                    auto reply = call.createErrorReply({dnfdaemon::ERROR, ex.what()});
-                    try {
-                        reply.send();
-                    } catch (const std::exception & e) {
-                        auto & logger = *base->get_logger();
-                        logger.error(fmt::format("Error sending d-bus error reply: {}", e.what()));
-                    }
-                }
-                threads_manager.current_thread_finished();
-            },
-            std::move(call));
-        threads_manager.register_thread(std::move(worker));
-    }
-
 private:
     sdbus::IConnection & connection;
     std::unique_ptr<libdnf::Base> base;
